+++ src/components/index.ts
@@ -1,3 +1,3 @@
 export { default as Aside } from './Aside/Aside';
 export { default as Header } from './Header/Header';
-export { default as Headings } from './Headings/Headings';
+export { default as Headings } from '@/components/Heading/Heading.tsx';
